requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-oauth-toolkit==1.2.0
djangorestframework==3.9.2
idna==2.8
oauthlib==3.0.1
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
